requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
dj-rest-auth==5.0.2
drf-spectacular==0.26.5
requests==2.31.0
psycopg2-binary==2.9.9
urllib3==2.1.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
djangorestframework_simplejwt==5.3.0
httpx==0.25.2